Apply improved types & architecture to Stacks
