+++ libs/ledger-live-common/src/families/stacks/bridge/account.ts
@@ -1,302 +1,17 @@
-import BlockstackApp from "@zondax/ledger-stacks";
-import { c32address } from "c32check";
-import { StacksMainnet } from "@stacks/network";
-import { BigNumber } from "bignumber.js";
-import { Observable } from "rxjs";
-import invariant from "invariant";
-import {
-  AmountRequired,
-  FeeNotLoaded,
-  InvalidAddress,
-  InvalidNonce,
-  InvalidAddressBecauseDestinationIsAlsoSource,
-  NotEnoughBalance,
-  RecipientRequired,
-} from "@ledgerhq/errors";
-import { getAbandonSeedAddress } from "@ledgerhq/cryptoassets";
-import { StacksMemoTooLong } from "../errors";
-import {
-  AddressVersion,
-  TransactionVersion,
-  AnchorMode,
-  UnsignedTokenTransferOptions,
-  estimateTransaction,
-  makeUnsignedSTXTokenTransfer,
-} from "@stacks/transactions";
-
-import { makeAccountBridgeReceive, makeSync } from "../../../bridge/jsHelpers";
+import { AccountBridge } from "@ledgerhq/types-live";
 import { defaultUpdateTransaction } from "@ledgerhq/coin-framework/bridge/jsHelpers";
-import {
-  Account,
-  AccountBridge,
-  AccountLike,
-  BroadcastFnSignature,
-  SignOperationEvent,
-  SignOperationFnSignature,
-} from "@ledgerhq/types-live";
-
-import { StacksNetwork } from "./utils/api.types";
-import { StacksOperation, Transaction, TransactionStatus } from "../types";
-import { findNextNonce, getAccountShape, getTxToBroadcast } from "./utils/misc";
-import { broadcastTx } from "./utils/api";
-import { getAddress } from "../../filecoin/bridge/utils/utils";
-import { withDevice } from "../../../hw/deviceAccess";
-import { getPath, throwIfError } from "../utils";
-import { getMainAccount } from "../../../account";
-import { encodeOperationId, patchOperationWithHash } from "../../../operation";
-import { validateAddress } from "./utils/addresses";
+import { makeAccountBridgeReceive, makeSync } from "../../../bridge/jsHelpers";
+import { getTransactionStatus } from "../getTransactionStatus";
+import { estimateMaxSpendable } from "../estimateMaxSpendable";
+import { prepareTransaction } from "../prepareTransaction";
+import { createTransaction } from "../createTransaction";
+import { signOperation } from "../signOperation";
+import { getAccountShape } from "./utils/misc";
+import { broadcast } from "../broadcast";
+import { Transaction } from "../types";
 
 const receive = makeAccountBridgeReceive();
-
-const createTransaction = (): Transaction => {
-  return {
-    family: "stacks",
-    recipient: "",
-    amount: new BigNumber(0),
-    network: "mainnet",
-    anchorMode: AnchorMode.Any,
-    useAllAmount: false,
-  };
-};
-
 const sync = makeSync({ getAccountShape });
-
-const broadcast: BroadcastFnSignature = async ({
-  signedOperation: { operation, signature, rawData },
-}) => {
-  invariant(operation as StacksOperation, "StacksOperation expected");
-  const tx = await getTxToBroadcast(operation as StacksOperation, signature, rawData ?? {});
-
-  const hash = await broadcastTx(tx);
-  const result = patchOperationWithHash(operation, hash);
-
-  return result;
-};
-
-const getTransactionStatus = async (a: Account, t: Transaction): Promise<TransactionStatus> => {
-  const errors: TransactionStatus["errors"] = {};
-  const warnings: TransactionStatus["warnings"] = {};
-
-  const { spendableBalance } = a;
-  const { address } = getAddress(a);
-  const { recipient, useAllAmount, fee } = t;
-  const { memo } = t;
-  let { amount } = t;
-
-  if (!recipient) {
-    errors.recipient = new RecipientRequired();
-  } else if (!validateAddress(recipient).isValid) {
-    errors.recipient = new InvalidAddress("", {
-      currencyName: a.currency.name,
-    });
-  } else if (address === recipient) {
-    errors.recipient = new InvalidAddressBecauseDestinationIsAlsoSource();
-  } else if (!fee || fee.eq(0)) {
-    errors.gas = new FeeNotLoaded();
-  }
-
-  const estimatedFees = fee || new BigNumber(0);
-
-  const totalSpent = useAllAmount ? spendableBalance : amount.plus(estimatedFees);
-  amount = useAllAmount ? spendableBalance.minus(estimatedFees) : amount;
-
-  if (amount.lte(0)) errors.amount = new AmountRequired();
-  if (totalSpent.gt(spendableBalance)) errors.amount = new NotEnoughBalance();
-  if (memo && memo.length > 34) errors.transaction = new StacksMemoTooLong();
-
-  return {
-    errors,
-    warnings,
-    estimatedFees,
-    amount,
-    totalSpent,
-  };
-};
-
-const estimateMaxSpendable = async ({
-  account,
-  parentAccount,
-  transaction,
-}: {
-  account: AccountLike;
-  parentAccount?: Account | null | undefined;
-  transaction?: Transaction | null | undefined;
-}): Promise<BigNumber> => {
-  const a = getMainAccount(account, parentAccount);
-  const { spendableBalance } = a;
-  const mainAccount = getMainAccount(account, parentAccount);
-  const { xpub } = mainAccount;
-  invariant(xpub, "xpub is required");
-
-  const dummyTx = {
-    ...createTransaction(),
-    ...transaction,
-    recipient: getAbandonSeedAddress(a.currency.id),
-    useAllAmount: true,
-  };
-  // Compute fees
-  const { recipient, anchorMode, memo, amount } = dummyTx;
-  const network = StacksNetwork[dummyTx.network] || new StacksMainnet();
-
-  const options: UnsignedTokenTransferOptions = {
-    recipient,
-    anchorMode,
-    memo,
-    network,
-    publicKey: xpub,
-    amount: amount.toFixed(),
-  };
-
-  const tx = await makeUnsignedSTXTokenTransfer(options);
-
-  const [feeEst] = await estimateTransaction(tx.payload);
-
-  const diff = spendableBalance.minus(new BigNumber(feeEst.fee));
-  return diff.gte(0) ? diff : new BigNumber(0);
-};
-
-const prepareTransaction = async (a: Account, t: Transaction): Promise<Transaction> => {
-  const { spendableBalance, pendingOperations } = a;
-  const { recipient, useAllAmount } = t;
-  const { xpub } = a;
-  invariant(xpub, "xpub is required");
-
-  const patch: Partial<Transaction> = {};
-  if (xpub && recipient && validateAddress(recipient).isValid) {
-    const { anchorMode, memo, amount } = t;
-
-    const network = StacksNetwork[t.network] || new StacksMainnet();
-
-    // Check if recipient is valid
-    const options: UnsignedTokenTransferOptions = {
-      recipient,
-      anchorMode,
-      memo,
-      network,
-      publicKey: xpub,
-      amount: amount.toFixed(),
-    };
-
-    const tx = await makeUnsignedSTXTokenTransfer(options);
-
-    const addressVersion =
-      network.version === TransactionVersion.Mainnet
-        ? AddressVersion.MainnetSingleSig
-        : AddressVersion.TestnetSingleSig;
-    const senderAddress = c32address(addressVersion, tx.auth.spendingCondition!.signer);
-
-    const [fee] = await estimateTransaction(tx.payload);
-
-    patch.fee = new BigNumber(fee.fee);
-    patch.nonce = await findNextNonce(senderAddress, pendingOperations);
-
-    if (useAllAmount) patch.amount = spendableBalance.minus(patch.fee);
-  }
-
-  return defaultUpdateTransaction(t, patch);
-};
-
-const signOperation: SignOperationFnSignature<Transaction> = ({
-  account,
-  deviceId,
-  transaction,
-}): Observable<SignOperationEvent> =>
-  withDevice(deviceId)(
-    transport =>
-      new Observable(o => {
-        async function main() {
-          const { address, derivationPath } = getAddress(account);
-          const { xpub, id: accountId } = account;
-          invariant(xpub, "xpub is required");
-
-          const { recipient, fee, anchorMode, network, memo, amount, nonce } = transaction;
-
-          if (!xpub) {
-            throw new InvalidAddress("", {
-              currencyName: account.currency.name,
-            });
-          }
-
-          if (!fee) {
-            throw new FeeNotLoaded();
-          }
-
-          if (!nonce) {
-            throw new InvalidNonce();
-          }
-
-          const blockstack = new BlockstackApp(transport);
-
-          const options: UnsignedTokenTransferOptions = {
-            amount: amount.toFixed(),
-            recipient,
-            anchorMode,
-            network: StacksNetwork[network],
-            memo,
-            publicKey: xpub,
-            fee: fee.toFixed(),
-            nonce: nonce.toFixed(),
-          };
-
-          const tx = await makeUnsignedSTXTokenTransfer(options);
-
-          const serializedTx = tx.serialize();
-
-          o.next({
-            type: "device-signature-requested",
-          });
-
-          // Sign by device
-          const result = await blockstack.sign(getPath(derivationPath), Buffer.from(serializedTx));
-          throwIfError(result);
-
-          o.next({
-            type: "device-signature-granted",
-          });
-
-          const txHash = "";
-
-          // build signature on the correct format
-          const signature = `${result.signatureVRS.toString("hex")}`;
-
-          const operation: StacksOperation = {
-            id: encodeOperationId(accountId, txHash, "OUT"),
-            hash: txHash,
-            type: "OUT",
-            senders: [address],
-            recipients: [recipient],
-            accountId,
-            value: amount.plus(fee),
-            fee,
-            blockHash: null,
-            blockHeight: null,
-            date: new Date(),
-            transactionSequenceNumber: nonce.toNumber(),
-            extra: {
-              memo,
-            },
-          };
-
-          o.next({
-            type: "signed",
-            signedOperation: {
-              operation,
-              signature,
-              rawData: {
-                xpub,
-                network,
-                anchorMode,
-              },
-            },
-          });
-        }
-
-        main().then(
-          () => o.complete(),
-          e => o.error(e),
-        );
-      }),
-  );
 
 export const accountBridge: AccountBridge<Transaction> = {
   createTransaction,
